test(preprocessor): simplify script contract coverage
