+++ tests/run_test_suite.py
@@ -79,12 +79,17 @@ UNIT_MODULES = frozenset(
         "test_ida_analyze_util",
         "test_ida_llm_utils",
         "test_ida_mcp_session",
-        "test_ida_preprocessor_scripts",
+        "test_ida_preprocessor_contracts_commands",
+        "test_ida_preprocessor_contracts_gamesystems",
+        "test_ida_preprocessor_contracts_misc",
+        "test_ida_preprocessor_contracts_network",
         "test_ida_remote_export",
+        "test_ida_skill_preprocessor",
         "test_ida_vcall_finder",
         "test_igamesystem_dispatch_common",
         "test_igamesystem_slot_dispatch_preprocessor",
         "test_igamesystem_slot_dispatch_py_eval_behavior",
+        "test_indirect_vcall_target_common",
         "test_init_gamebin",
         "test_llm_decompile_dependencies",
         "test_process_api",
@@ -94,6 +99,7 @@ UNIT_MODULES = frozenset(
         "test_process_scheduler_redis",
         "test_process_status_reader_redis",
         "test_prune_pr_expected_output_bin",
+        "test_ordinal_vtable_common",
         "test_register_api_callbacks_preprocessor",
         "test_register_event_listener_abstract_preprocessor",
         "test_registerconcommand_preprocessor",
